# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação",
        "description": "API para gerenciamento de autenticação de usuários",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/profile": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Exibe o perfil do usuário autenticado",
                "description": "Retorna os detalhes do perfil do usuário autenticado. Requer autenticação.",
                "operationId": "aeb0a04ac557b5598c94ba676b75d809",
                "responses": {
                    "200": {
                        "description": "Perfil retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "john@example.com"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Desenvolvedor e entusiasta de tecnologia.",
                                                    "nullable": true
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "example": "avatars/1/profile.jpg",
                                                    "nullable": true
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "editor"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-05T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-05T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao carregar perfil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao carregar perfil."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Atualiza o perfil do usuário autenticado",
                "description": "Atualiza os dados do perfil do usuário autenticado, incluindo nome, username, email, bio e avatar (apenas para editores e administradores). Requer autenticação. Suporta envio de dados como JSON ou multipart/form-data.",
                "operationId": "f002c5f576ef466b1efaece2707fd162",
                "requestBody": {
                    "description": "Dados do perfil a serem atualizados. Escolha entre JSON ou multipart/form-data (necessário para upload de arquivo).",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "description": "Nome de usuário único",
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "description": "Email único",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john@example.com"
                                    },
                                    "bio": {
                                        "description": "Biografia do usuário",
                                        "type": "string",
                                        "maxLength": 5000,
                                        "example": "Desenvolvedor e entusiasta de tecnologia.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "description": "Nome de usuário único",
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "description": "Email único",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "john@example.com"
                                    },
                                    "bio": {
                                        "description": "Biografia do usuário",
                                        "type": "string",
                                        "maxLength": 5000,
                                        "example": "Desenvolvedor e entusiasta de tecnologia.",
                                        "nullable": true
                                    },
                                    "avatar": {
                                        "description": "Imagem do avatar (JPEG, PNG, JPG, GIF, WEBP, máx: 2MB). Obrigatório para editores/administradores se enviado, opcional para outros.",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil atualizado com sucesso."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "john@example.com"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Desenvolvedor e entusiasta de tecnologia.",
                                                    "nullable": true
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "example": "avatars/1/profile.jpg",
                                                    "nullable": true
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "editor"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-05T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-05T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não autorizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Você não tem permissão para fazer upload de avatar."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The avatar field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "avatar": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The avatar is required for editors and administrators."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar perfil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao atualizar perfil. Tente novamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the Bearer token obtained from the login or register endpoint to access protected routes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}